Remove unnecessary lazy loading
